add second arg to res.render to pass data into template
we want to add to the errors array in User.js from our session data. We created an object with the key error whose value comes from req.flash('errors'). 'errors'. The string 'errors' is not the same thing as the actual array called errors. The string 'errors' is just the name of the value. We could access that data manually by typing req.session.flash.errors but using the flash('errors'), we can delete it from the session so that the user only sees the message once