create error message if username already exists
first create if statement to check if username is actually valid. Once username proves valid, create a variable called usernameExists. usernameExists is assigned by taking the usersCollection from mongodb and finding  the first document where the username matches what the user tried to register with (this.data.username). If mongodb does find a matching document, then we will want to display the error message "That username is already taken". If username does not exist, the object will evaluate to null and allow user to sign up. *first make the entire function a promise (async/await) to ensure the correct timing of the if statements for mongodb.