:recycle: replace findOne with aggregate
aggregate is used to perform multiple operations, in this case $match & $lookup. This is done in an array with the aggregate function/method. We first are matching the id with mongoDB ObjectID method, and then looking up the 'users' collection, finding the author, and matching the id field. as: authorDocument is used by mongoDB when it adds on a virtual field or property with the matching user document to the post. authorDocument is displayed in the console as another property within the postCollection part of the database