:sparkles: add 3rd aggregate operation, $project
project allows us to spell out exactly what fields we want the resulting object to have. this way we have control on what we want to return. (title,body,createdDate, author). the 1 represents true. for the author field, we want the entire authorDocument. $arrayElemAt brings in an array that holds the authorDocument that was found during the $lookup. mongoDB sees $ and knows to pull in the entire document, not just a string of text. since there's only ever 1 authorDocument, we are returning the first item [0] in the array.