:sparkles: create logic for ifUserExists
create a function within called findByUsername which takes the req and looks at the params and grabs the username. if it resolves by finding the username, take that data as a new property on the request object (profileUser) and assign it to userDocument so that the next() function for the route can access it.