:sparkles: create logic for findByUsername in the models folder for User
the User function will call findByUsername. this function first returns a promise. if the type of username is NOT a string, then reject. If it is a string of text, go into the database in the usersCollection and find the username document. if the promise was successful THEN pass in that param of data called userDoc. if that param of data, userDoc exists, set userDoc to equal a new instance of User. what we want to get back from the database is an object that has the _id, username, and avatar. if the database has an error, reject(). Once it grabs that userDoc data, resolve the promise with the updated value of userDoc so that we can use to display the username, avatar, etc. if the promise is unsuccessful, CATCH and reject()