:sparkles: create property, posts to add to the userProfile object
findByAuthorId takes in the profileUser data to grab the id data. we need to go to the database and find a post by the author matching the id. here we just set up a promise to receive the posts or send an error if there is a technical issue. we will create the actual logic for the findByAuthorId in the Post models file