:recycle: create reusable function to connect to the database
findSingleById is refactored to use the reusable function to get specific data from the database. Then we created the function, findAuthorById which also uses the reusable function to connect to the database.  here, we match the author to the authorId and sort from newest post. -1 is used because its for descending order