add code that runs at the start of every req 2 check if visitor has id
if they have an id, we can grab that, if they don't, their id will be 0 and 0 will signify that they are a guest. to set up this code that runs for every request, we needed to set it up in the main app file. within app.js we already have a function that runs for every req. it takes the session req and makes it available for res.locals meaning we can access it from our ejs templates. so we just added additional code make current user id available on the req object.