:sparkles: match authorId to visitorId in database
postController is calling the function findSingleById. this function is leveraging our reusablePostQuery function. we add a second param to our findSingleById function to receive the incoming visitorId. then when we use our reusable function, the first arg is an array of aggregate functions, and the second arg will be visitorId. we go into the reusable function and leverage the visitorId first by receiving it as a second param. then when we map on each post, we can add a new property isVisitor owner to check to see if the post author id matches that passed in visitorId. authorId's value '$author' is a mongodb method to get the actual field and not a string of text.