:sparkles: create function to edit post
we want to make a new instance of the post that takes in the req.body, req.visitorId, and req.params.id. the post has a function inside called update (that we will make in the Post models file) that updates in the post was sucessfully updated in the DB. We send a success message via flash and redirect to the edit page again. If not successful (validation errors), we will send an error message via flash and redirect back to the edit page. We also catch any errors if a post id doesn't exist or if the visitor's id doesn't match. we save that session and redirect to the homepage