:sparkles: create function to update post in mongoDB
first create function to update. this function returns a new promise that either resolves or rejects. we have a try/catch set up. in the try portion we have made  post to equal the function, findSingleById, which takes in two args requestedPostId and userId (this data comes from the middleware postController through Post constructor). if the post is the owner ( if the authorId matches visitorId ) then called the actuallyUpdate(). We then create the actuallyUpdate() that takes in a new promise. this just makes sure there's no errors by calling the cleanUp and validate functions. if no errors go to the database and find the id that matches the newly requestedPostId and set the changes to the title and body. we resolve with a success status. else resolve with a failure status. in the update function, the status is then sent back to the middleware to use