:sparkles: assign an id to the newly created post
in our Post model file, we go to our create method. when the db inserts the data with insertOne, it sends back a ton of information, including a new id. within the then() we add a param called info, so when we resolve, we resolve  with a value. this value looks inside the info, mongoDB has an array called ops and we only want the first element in that array [0] which then we look inside that to grab the _id. now our promise revolves with that brand new, newly created id. then we go back in the postController, pass in the newId as a param and add that newId to the route