:sparkles: middleware for delete function that will be in Post model
this is setting up the middleware for the actual delete function that will be created next in the Post model. what's happening in this code is the bridge between the route and the model. first we have a function called delete. this delete function is the shell of the delete function that will be in Post model. here we take in the req, res from the Promise that will be in the Post models delete function.   the second delete function is the Post models delete function that is bringing in the id's as an argument. if the req is a success, send message and send user back to their profile page. if the req has an error, redirect back to the home page.