:sparkles: create actual delete function that connects to the database
this function takes in two params, postIdToDelete and currentUser. we create a promise that will resolve or reject. we set post to equal a post that uses findSingleById function. this function gets called and takes in the id and visitorId as args (postIdToDelete and currentUser). if the post is owned by the visitor, we await for the database to delete one post where the _id key uses the mongoDB object method to set the postIdToDelete to be the value. resolve() is called. else reject()