:construction: set up settings for webpack config
when we leverage webpack it looks for this file. in this file we tell webpack where our files live, where it should export or bundle the files to (public folder), and any additional features we'd want it to use. essentially we are telling webpack where our desired JS files live ('./frontend-js/main.js'), where we want to export them to ('public'), and setting up the use of babel. babel lets us write ultra modern javascript code in syntax and converts it into javascript that can be supported by almost every web browser.