:sparkles: add an alert message when search bar icon is clicked
set up 4 steps. 1. is to select DOM elements within the constructor In this case we are selecting the header-search-icon class. 2. set up the event for this element. we add an event listener, click and prevent default. then we call the actual method 3. openOverlay is the actual method that has the alert. in order for the final result of the alert popping up, we need to do step 4. call the events function.