:sparkles: create method to display search HTML
this is just the beginning of what we actually want out of this function. but for now we created a function injectHTML. inside the function we grab the document and inside the document we to to the body portion. then we call a method insertAdjacentHTML which takes in two args. the first argument is where we want to insert the HTML ('beforeend') and the 2nd arg is the actual HTML file using back ticks. finally we call this method in our class. we place this above everything because we want the user to see the HTML page