:sparkles: display the search-overlay when search icon is clicked
before in the HTML we had the class search-overlay--visible. this made that HTML file automatically display. we only wanted the file to display when the search icon is clicked. so we selected the DOM element search-overlay and set it to a variable, overlay. then in our openOverlay function we used our overlay (that selects the DOM element), went inside the class list to use a method called add. so when the search icon is clicked, add() will add the search-overlay--visible to the classList which then displays the search HTML file