:sparkles: close the search-overlay when close icon is clicked
first we created a variable called closeIcon that takes the document and selects the class close-live-search. then we added the closeIcon to our events by adding an event listener that calls our closeOverlay function when the icon gets clicked. when closeOverLay function is called, the overlay goes into the classList and removes the search-overlay--visible class