:sparkles: show spinning loader icon when user types in the input field
first we add a variable to our constructor called loaderIcon. the document uses querySelector to grab the class circle-icon. we then need to create an event that when someone types in the inputField, 'keyup' eventListener, it triggers a function called keyPressHandler(). this function is a method that triggers a function called showLoaderIcon(). showLoaderIcon() selects the classList which uses the method add() to add the class circle-loader--visible. so when the user starts typing, the loader icon shows up.