:sparkles: create beginning of function to search the users input value
first we create two variables, typingWaitTimer (which does not get set to a value just yet) and previousValue (value starts off as an empty string). then in our keyPressHandler(), instead of just running the loader icon, we need to add logic so that it displays at the right time. we create a variable called value which takes the value from what's typed in the inputField. we have an if statement that says if the value doesn't equal an empty string and the value doesn't equal the previousValue (meaning changes were made in the inputField), then we can display the loader icon. we also use our typingWaitTimer variable to equal a setTimeout  that calls a method called sendRequest(). we use the setTimeout function because we don't want to send an actual request to the database until the user is done typing. for example if someone was searching "puppy" we wouldnt want it to send a request to the letter "p", then "pu", then "pup". we'd want "puppy" to be sent over. right now the sendRequest() just sends us an alert that it's working