:sparkles: leverage axios to perform a post request to the backend
we use axios and it's post method to send a post request. this method takes in two arguments, the first is a URL that we want to send the request to and the second arg is an object of data that we want to send along to the server. we created a property called searchTerm who's value comes from the inputField's value. the axios library is going to return a promise so we add .then() and .catch(). for now we don't have the route created on our backend, so to test it out, we just created an alert message in the .catch()