:construction: :sparkles: set up search route in the backend
we create a post method on our router that takes in the route for the first arg, and the postController with the method of search as our second arg, Inside our middleware, postController, we create the search function. this function takes in params req and res. Inside the postController search function, we use our Post model to use it's own search method that takes the req.body.searchTerm which is the object that came along for the ride when we used axios on the frontend. The Post.search method is running a Promise (logic is in the Post model for this promise) AFTER this Promise is resolved in the Post models is when we either THEN or CATCH. In the then(), if the promise is successful and resolves, we just want to send back the raw json data to the browser. we set this up so that the promise resolves with a value. the value will be the data, (posts),  from the database that it finds. if there is an error or the promise rejects for any reason, the catch will send back an empty array