:sparkles: Create the Promise logic for Post.search
Post.search method takes in the searchTerm and creates a new Promise around it. If the type of searchTerm is a string, then we perform the database methods. so we create a variable called posts which takes Post and called the reusablePostQuery function. reminder that reusablePostQuery function runs the aggregate boilerplate code. once that code runs, we use the bd method $match to match any text within our searchTerms value. we use the db method $sort to sort by score which means relevant score (best match for the users searchTerm). else reject()
**note for what we did in the Database to prevent this operation to cost us: in the database we clicked collects, indexes and added an index. "title": "text" "body": "text" which outputs title_text_body_text for a lighter and easier search