:sparkles: display correct HTML for the live search results
our sendRequest() will take the response.data from the database and pass as a param in a new function called renderResultsHMTL(). when this function is called. it displays the correct HTML depending on whats searched. If there are posts in the array (posts.length), display the posts by taking the resultsArea and setting it's innerHTML to the correct HTML. (we took the live search results HTML from the injectHTML() function.) else display the HTML as a paragraph with a "Sorry" message.