:sparkles:create methods to hide loader icon & display/hide live results
created 3 new methods. hideLoaderIcon() removes the class 'circle-loader--visible' from the classList in our HTML. showResults() adds the class 'live-search-results--visible' to the classList which displays the search results. hideResultsArea() removes 'live-search-results--visible' from the classList. 

placed showLoaderIcon() and hideResultsArea() functions in when user is typing into the search bar. here the loader icon will be firing while the database is getting the results. created a new if statement that states if the value of the search bar is an empty string, then we will stop the loader icon and hide the results area (this is what the user will see first when clicking on the search area)