dynamically add data from posts through mapping
display the number of posts found with a ternary; create a variable called postDate that takes a new instance of Date() and takes in the mapped post which holds the createdDate data. we dynamically pull in the rest of the data and place in appropriate areas. at the end of this map function, we call the join method to bring it call together